Addresses sources of UB in generator
- Enables lots of warnings with -Wall -Wextra -Wpedantic
- Makes warning into errors with -Werror
  - no-strict-aliasing for now
- Enables undefined behavior sanitizer

This fixes the existing warnings

[ROCm/mxDataGenerator commit: 9be95799266dc55f9eaa4a71a84ca2327d432506]

[ROCm/mxDataGenerator commit: 32bbbf6530b4866f7b8f380d8fb5bd822f1026b6]

+++ shared/mxdatagenerator/shared/mxdatagenerator/test/bf16_test.cpp
@@ -1192,8 +1192,8 @@ TEST_F(bf16_test, getDataMin)
 TEST_F(bf16_test, getDataMaxSubnorm)
 {
     float subMax = 0;
-    for(int i = 0; i < getDataMantissaBits<DT>(); i++)
-        subMax += std::pow(2, -(i + 1));
+    for(size_t i = 0; i < getDataMantissaBits<DT>(); i++)
+        subMax += std::pow(2, -int32_t(i + 1));
 
     subMax *= std::pow(2, -126);
 
@@ -1495,7 +1495,7 @@ TEST_F(bf16_test, preserveSign)
 
     t.num = std::numeric_limits<float>::quiet_NaN();
 
-    t.bRep |= 1 << 31;
+    t.bRep |= 1U << 31;
 
     uint16_t sOutput  = satConvertToType<DT>(t.num);
     uint16_t nSOutput = nonSatConvertToType<DT>(t.num);
